Larger Precision Calculation

+++ app/src/main/kotlin/com/simplemobiletools/calculator/operation/base/Operation.kt
@@ -1,5 +1,7 @@
 package com.simplemobiletools.calculator.operation.base
 
+import java.math.BigDecimal
+
 interface Operation {
-    fun getResult(): Double
+    fun getResult(): BigDecimal
 }
